scripts/requirements.txt
zhipuai
pygithub
